Added Mocha and Sinon framework & spy support
Fixes #4.

This is a rather large change. Instead of directly calling `afterEach` and so on by the detected environment, the environment is treated as a set of "hooks" used by the `cft` driver. This is necessary because different hooks might need to be called by different environments: `before`/`beforeEach` vs. `beforeAll`/`before`, for example.

The spy library in use is now also decoupled from the test framework. Mocha sometimes is run with nothing and sometimes is run with Sinon.